Managed to export an economy.html file.
I haven't worked out how to actually produce usable code from it yet.
I hope to be able to produce a Jinja enabled template that I can then render.
This way it should be easy for elthran to add features but allow me to slowly convert to full Vue stlye.