Add redirect to login on 401 error in Vue request.
Add 404 component when you got to a url that doesn't exist.
I'm still not entirely sure what will happen if you have the page open and your county gets deleted.
Add a general url builder API.
Add an advance_age script, though I'm not sure it will work on the server.
Attempt to add a hook that redirects you to initialization page if you are logged in but have no county.